api/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
google-cloud-texttospeech==2.16.3
gunicorn
